Move mapProject files into MapProject folder
