+++ BabelWiresLib/CMakeLists.txt
@@ -39,14 +39,14 @@ SET( BABELWIRESLIB_SRCS
 	Types/Map/Commands/changeEntryKindCommand.cpp
 	Types/Map/Commands/setMapCommand.cpp
 	Types/Map/mapFeature.cpp
-	Types/Map/mapProject.cpp
-	Types/Map/mapSerialization.cpp
 	Types/Map/mapValue.cpp
 	Types/Map/MapEntries/mapEntryData.cpp
 	Types/Map/MapEntries/oneToOneMapEntryData.cpp
 	Types/Map/MapEntries/allToSameFallbackMapEntryData.cpp
 	Types/Map/MapEntries/allToOneFallbackMapEntryData.cpp
-	Types/Map/mapProjectEntry.cpp
+	Types/Map/MapProject/mapProject.cpp
+	Types/Map/MapProject/mapProjectEntry.cpp
+	Types/Map/MapProject/mapSerialization.cpp
 	Types/Map/mapType.cpp
 	Types/Map/mapTypeConstructor.cpp
 	Processors/parallelProcessor.cpp
