+++ BabelWiresLib/Types/Map/MapProject/mapProject.cpp
@@ -5,10 +5,10 @@
  *
  * Licensed under the GPLv3.0. See LICENSE file.
  **/
-#include <BabelWiresLib/Types/Map/mapProject.hpp>
+#include <BabelWiresLib/Types/Map/MapProject/mapProject.hpp>
 
 #include <BabelWiresLib/Types/Map/MapEntries/mapEntryData.hpp>
-#include <BabelWiresLib/Types/Map/mapProjectEntry.hpp>
+#include <BabelWiresLib/Types/Map/MapProject/mapProjectEntry.hpp>
 #include <BabelWiresLib/Project/projectContext.hpp>
 #include <BabelWiresLib/Types/Int/intType.hpp>
