+++ BabelWiresQtUi/ComplexValueEditors/MapEditor/MapEntryModels/mapEntryModel.cpp
@@ -14,7 +14,7 @@
 #include <BabelWiresLib/Types/Map/MapEntries/allToOneFallbackMapEntryData.hpp>
 #include <BabelWiresLib/Types/Map/MapEntries/allToSameFallbackMapEntryData.hpp>
 #include <BabelWiresLib/Types/Map/MapEntries/oneToOneMapEntryData.hpp>
-#include <BabelWiresLib/Types/Map/mapProjectEntry.hpp>
+#include <BabelWiresLib/Types/Map/MapProject/mapProjectEntry.hpp>
 
 void babelwires::MapEntryModel::init(const ValueModelRegistry& valueModelRegistry) {}
